+++ shuup/core/api/__init__.py
@@ -5,9 +5,10 @@
 #
 # This source code is licensed under the OSL-3.0 license found in the
 # LICENSE file in the root directory of this source tree.
+from shuup.core.api.address import MutableAddressViewSet
 from shuup.core.api.attribute import AttributeViewSet
 from shuup.core.api.category import CategoryViewSet
-from shuup.core.api.contacts import ContactViewSet
+from shuup.core.api.contacts import ContactViewSet, PersonContactViewSet
 from shuup.core.api.manufacturer import ManufacturerViewSet
 from shuup.core.api.orders import OrderViewSet
 from shuup.core.api.product_media import ProductMediaViewSet
@@ -24,6 +25,7 @@ from shuup.core.api.suppliers import SupplierViewSet
 from shuup.core.api.tax_class import TaxClassViewSet
 from shuup.core.api.units import SalesUnitViewSet
 from shuup.core.api.users import UserViewSet
+from shuup.core.api.basket import BasketViewSet
 
 
 def populate_core_api(router):
@@ -31,10 +33,12 @@ def populate_core_api(router):
     :param router: Router
     :type router: rest_framework.routers.DefaultRouter
     """
+    router.register("shuup/address", MutableAddressViewSet)
     router.register("shuup/attribute", AttributeViewSet)
     router.register("shuup/category", CategoryViewSet)
     router.register("shuup/contact", ContactViewSet)
     router.register("shuup/order", OrderViewSet)
+    router.register("shuup/person_contact", PersonContactViewSet)
     router.register("shuup/product", ProductViewSet)
     router.register("shuup/product_attribute", ProductAttributeViewSet)
     router.register("shuup/product_media", ProductMediaViewSet)
@@ -50,3 +54,4 @@ def populate_core_api(router):
     router.register("shuup/user", UserViewSet)
     router.register("shuup/sales_unit", SalesUnitViewSet)
     router.register("shuup/tax_class", TaxClassViewSet)
+    router.register("shuup/basket", BasketViewSet)