+++ doc/changelog.rst
@@ -15,6 +15,9 @@ Core
 - Add `created_on` and `modified_on` fields for shop
 - Make shop identifier max length to 128 characters
 - Add `staff_members` manytomanyfield for shop
+- API: add address endpoint
+- API: add person contact endpoint
+- API: add user register endpoint
 
 Localization
 ~~~~~~~~~~~~